#9 add initial logic for code evaluation
codeRunner.js creates a python file from variables (these will be replaced with actual testCases from the db as well as user input code) and then runs the code and checks the testCases. Right now we need to figure out how to ignore eventual print statements inside the user code, as well as put it all in a container.