#2 add pinia user store and auth/me endpoint
This implements a global user store in the frontend. When the page loads, the cookie is used to get the user from the backend. Info is then stored. The intention is that, if the cookie is not there or it doesn't return a user from the backend, the page will show the logged out view. This can be checked by accessing the user.loggedIn attribute of the user store.